% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phytagoras.R
\name{Phytagoras}
\alias{Phytagoras}
\title{Pythagorean theorem}
\usage{
Phytagoras(a, b)
}
\description{
Pythagorean theorem is the length of two sides at right angles in a right triangle is a and b,
and if the length of the rain side is c, a^2+b^2=c^2 is established.
}
\examples{

Phytagoras(3,4)
}
